Create README.md
A program that uses console.log to print all the number from 1-200, with two exceptions. For numbers divisible by 6, print "shams" instead of the number, and for numbers divisible by 8(and not 6) print "TIIDELAB" instead. When you have the working, modify your program to print "ShamsTIIDELAB" for number that are divisible by both 6 and 8.